requirements.txt
numpy
pandas
scikit-learn
sqlalchemy
pyodbc
jupyter
matplotlib
seaborn
missingno
python-dotenv
```